// src/types.ts
export interface TokenBasics {
    name: string;
    address: string
}


export interface DevMonitoredTokens extends TokenBasics {
    creator: string;
    creator_time: number;
    chain: string
}